+++ app/src/main/java/com/example/evolutionslek/IsDead.java
@@ -29,7 +29,7 @@ public class IsDead extends AppCompatActivity {
                 klar = true;
                 Toast.makeText(getApplicationContext(), Boolean.toString(klar), Toast.LENGTH_SHORT).show();
                 djur = data.getParcelableExtra("result");
-                String text = (Integer.toString(djur.mass) + "," + Integer.toString(djur.horns) + "," + Integer.toString(djur.speed) + "," + Integer.toString(djur.defense) + "," + Integer.toString(djur.maxHealth) + "," + Integer.toString(djur.claws) + "," + Integer.toString(djur.attack) + "," + Boolean.toString(djur.herbivore));
+                String text = (Double.toString(djur.mass) + "," + Double.toString(djur.horns) + "," + Double.toString(djur.speed) + "," + Double.toString(djur.defense) + "," + Double.toString(djur.maxHealth) + "," + Double.toString(djur.claws) + "," + Double.toString(djur.attack) + "," + Boolean.toString(djur.herbivore));
                 Toast.makeText(getApplicationContext(), text, Toast.LENGTH_SHORT).show();
 
                 Intent returnIntent = new Intent();
